Source: https://github.com/msupply-foundation/open-msupply/commit/ef8224acee9a55ecd0f9b00e5002ec44b1f9952d

+++ server/graphql/general/src/queries/tests/names.rs
@@ -146,7 +146,7 @@ mod graphql {
                   "nodes": [{
                       "id": mock_name_a().id,
                   }],
-                  "totalCount": 1 as i32
+                  "totalCount": 1_i32
               }
           }
         );
